+++ dpytools/errors.py
@@ -4,7 +4,7 @@ This Module contains all custom errors used in the package.
 """
 from discord.ext.commands import BadArgument, UserInputError
 
-from discord.ext.commands import CommandError
+from discord.ext.commands import CommandError, CheckFailure
 
 
 class NotMemberOfCorrectGuild(CommandError):
@@ -28,4 +28,7 @@ class InvalidTimeString(BadArgument):
 
 
 class MemberNorUserFound(UserInputError):
+    pass
+
+class OutsidePermittedDatetime(CheckFailure):
     pass